backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-filter==23.3
django-cors-headers==4.3.0
tqdm>=4.65.0
pysqlite3-binary>=0.5.0